feat(verifier): move verifier as a feature
