+++ Cargo.toml
@@ -6,18 +6,39 @@ edition = "2018"
 
 [dependencies]
 pairing_ce = { version = "0.24.2" }
-bls = { package = "bls12_381", version = "0.3.1" }
-bn = { git = "https://github.com/paritytech/bn", default-features = false }
-zkp-u256 = { git = "https://github.com/patractlabs/OpenZKP.git", version = "0.2.0", features = ["stable"] }
-num-bigint = "0.3"
-num-traits = "0.2"
 
+# For the `verifier` trait
+[dependencies.zkp-u256]
+git = "https://github.com/patractlabs/OpenZKP.git"
+version = "0.2.0"
+features = ["stable"]
+optional = true
+
+[dependencies.num-bigint]
+version = "0.3"
+optional = true
+
+[dependencies.num-traits]
+version = "0.2"
+optional = true
+
+# For tests
 [dev-dependencies]
 hex-literal = "0.2.1"
 rand = { version = "0.5", features = ["i128_support"], default-features = false }
 rustc-hex = { version = "2", default-features = false }
 
+# Local CI
 [dev-dependencies.cargo-husky]
 version = "1"
 features = ["user-hooks"]
 default-features = false
+
+# Features
+[features]
+default = []
+verifier = [
+         "num-bigint",
+         "num-traits",
+         "zkp-u256",
+]
