+++ src/scratch.rs
